server/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
amqp==2.6.1
    # via kombu
arrow==0.13.0
    # via
    #   eve-elastic
    #   superdesk-core
authlib==0.14.3
    # via superdesk-core
babel==2.10.3
    # via flask-babel
bcrypt==3.1.7
    # via superdesk-core
billiard==3.6.4.0
    # via celery
blinker==1.4
    # via
    #   elastic-apm
    #   flask-mail
    #   raven
    #   superdesk-core
boto3==1.24.62
    # via superdesk-core
botocore==1.27.62
    # via
    #   boto3
    #   s3transfer
cachetools==5.2.0
    # via flask-oidc-ex
celery[redis]==4.4.7
    # via superdesk-core
cerberus==1.3.4
    # via
    #   eve
    #   superdesk-core
certifi==2022.6.15
    # via
    #   elastic-apm
    #   elasticsearch
    #   requests
cffi==1.15.1
    # via
    #   bcrypt
    #   cryptography
chardet==3.0.4
    # via superdesk-core
charset-normalizer==2.1.1
    # via requests
ciso8601==1.0.8
    # via eve-elastic
click==7.1.2
    # via flask
croniter==0.3.37
    # via superdesk-core
cryptography==37.0.4
    # via
    #   authlib
    #   jwcrypto
deepdiff==3.3.0
    # via superdesk-planning
deprecated==1.2.13
    # via jwcrypto
draftjs-exporter[lxml]==2.1.7
    # via superdesk-core
elastic-apm[flask]==6.7.2
    # via superdesk-core
elasticsearch==7.13.4
    # via eve-elastic
eve==1.1.2
    # via superdesk-core
eve-elastic==7.2.1
    # via superdesk-core
events==0.3
    # via eve
feedparser==6.0.10
    # via superdesk-core
flask==1.1.4
    # via
    #   eve
    #   flask-babel
    #   flask-mail
    #   flask-oidc-ex
    #   flask-script
    #   raven
    #   superdesk-core
flask-babel==1.0.0
    # via superdesk-core
flask-mail==0.9.1
    # via superdesk-core
flask-oidc-ex==0.5.5
    # via superdesk-core
flask-script==2.0.6
    # via superdesk-core
future==0.18.2
    # via python-twitter
gspread==3.1.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
hachoir==3.0a3
    # via superdesk-core
hermescache==0.7.2
    # via superdesk-core
honcho==1.0.1
    # via -r requirements.in
httplib2==0.20.4
    # via oauth2client
icalendar==4.0.9
    # via superdesk-planning
idna==3.3
    # via requests
importlib-metadata==3.1.1
    # via superdesk-core
itsdangerous==1.1.0
    # via
    #   flask
    #   flask-oidc-ex
jinja2==2.11.3
    # via
    #   flask
    #   flask-babel
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonpickle==2.2.0
    # via deepdiff
jwcrypto==1.3.1
    # via
    #   flask-oidc-ex
    #   python-jwt
kombu==4.6.11
    # via
    #   celery
    #   superdesk-core
ldap3==2.5.2
    # via superdesk-core
lxml==4.6.5
    # via
    #   draftjs-exporter
    #   superdesk-core
markupsafe==2.0.1
    # via
    #   jinja2
    #   superdesk-core
mongolock==1.3.4
    # via superdesk-core
natsort==8.1.0
    # via croniter
oauth2client==4.1.3
    # via
    #   -r requirements.in
    #   flask-oidc-ex
oauthlib==3.2.0
    # via requests-oauthlib
pillow==8.1.2
    # via superdesk-core
pyasn1==0.4.8
    # via
    #   ldap3
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via oauth2client
pycparser==2.21
    # via cffi
pymongo==3.11.4
    # via
    #   eve
    #   mongolock
    #   superdesk-core
pyparsing==3.0.9
    # via httplib2
python-dateutil==2.7.5
    # via
    #   arrow
    #   botocore
    #   croniter
    #   icalendar
    #   superdesk-core
python-jwt==3.3.3
    # via flask-oidc-ex
python-magic==0.4.27
    # via superdesk-core
python-twitter==3.5
    # via superdesk-core
pytz==2022.2.1
    # via
    #   babel
    #   celery
    #   eve-elastic
    #   flask-babel
    #   icalendar
    #   superdesk-core
    #   tzlocal
pyyaml==5.4.1
    # via superdesk-core
raven[flask]==6.10.0
    # via superdesk-core
redis==3.2.1
    # via
    #   celery
    #   superdesk-core
regex==2020.7.14
    # via superdesk-core
requests==2.28.1
    # via
    #   gspread
    #   python-twitter
    #   requests-oauthlib
    #   superdesk-core
requests-oauthlib==1.3.1
    # via python-twitter
rsa==4.9
    # via oauth2client
s3transfer==0.6.0
    # via boto3
sgmllib3k==1.0.0
    # via feedparser
simplejson==3.17.6
    # via eve
six==1.16.0
    # via
    #   bcrypt
    #   flask-oidc-ex
    #   oauth2client
    #   python-dateutil
superdesk-analytics @ git+https://github.com/superdesk/superdesk-analytics.git@v2.5.0-rc1
    # via -r requirements.in
superdesk-core @ git+https://github.com/superdesk/superdesk-core.git@v2.5.0-rc4
    # via -r requirements.in
superdesk-planning @ git+https://github.com/superdesk/superdesk-planning.git@v2.5.0-rc4
    # via -r requirements.in
typing-extensions==4.3.0
    # via superdesk-core
tzlocal==2.1
    # via superdesk-core
unidecode==0.04.21
    # via superdesk-core
urllib3==1.25.11
    # via
    #   botocore
    #   elastic-apm
    #   elasticsearch
    #   requests
    #   superdesk-core
vine==1.3.0
    # via
    #   amqp
    #   celery
websockets==6.0
    # via superdesk-core
werkzeug==1.0.1
    # via
    #   flask
    #   superdesk-core
wrapt==1.14.1
    # via deprecated
zipp==3.8.1
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
